added build method for images and the render time went down from 1.5s to 25ms

+++ helium/src/widgets/vstack.rs
@@ -12,7 +12,7 @@ pub struct VStack {
     children: Vec<Box<dyn Widget>>,
     color: Color,
     layout: VerticalLayout,
-	corner_radius:u8
+	corner_radius:u32
 }
 
 impl VStack {
